SaveState: Move file generation logic out of GUI
